api [nfc]: Move RealmEmojiItem to model, from initial_snapshot
This type is used in events, too, so it goes best in the
"model" library that's shared between those.
