+++ lib/api/model/initial_snapshot.dart
@@ -125,34 +125,6 @@ class RealmDefaultExternalAccount {
   Map<String, dynamic> toJson() => _$RealmDefaultExternalAccountToJson(this);
 }
 
-/// An item in `realm_emoji`.
-///
-/// For docs, search for "realm_emoji:"
-/// in <https://zulip.com/api/register-queue>.
-@JsonSerializable(fieldRename: FieldRename.snake)
-class RealmEmojiItem {
-  final String id;
-  final String name;
-  final String sourceUrl;
-  final String? stillUrl;
-  final bool deactivated;
-  final int? authorId;
-
-  RealmEmojiItem({
-    required this.id,
-    required this.name,
-    required this.sourceUrl,
-    required this.stillUrl,
-    required this.deactivated,
-    required this.authorId,
-  });
-
-  factory RealmEmojiItem.fromJson(Map<String, dynamic> json) =>
-    _$RealmEmojiItemFromJson(json);
-
-  Map<String, dynamic> toJson() => _$RealmEmojiItemToJson(this);
-}
-
 /// An item in `recent_private_conversations`.
 ///
 /// For docs, search for "recent_private_conversations:"
